Take a major step forward in getting PyJuggler working on Windows.  It turns out that Visual C++ can't handle the intense use of templates when there are a bunch of boost::python::class_<> declarations in a single file, and the recommendation on the Boost.Python FAQ is to split the source into multiple files (or to upgrade the compiler).
This code has been regenerated with the latest version of Pyste from the
Boost CVS repository, and I have done my best to migrate all of my
customizations from the "monolithic" .cpp into the multi-file scenario we
have now.  I have tested the Python and C++ example applications as well
as PySim from the VR Juggler Toolbox.  To the best of my knowledge, all
functionality has been retained, and the code does compile on Windows.
Getting it to run would be the next step.  :)

Prior to this change, the source tree was tagged with the symbol
PRE_FILE_EXPLOSION.  After this commit, the tree will be tagged again with
the symbol POST_FILE_EXPLOSION.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/PyJuggler@12985 08b38cba-cd3b-11de-854e-f91c5b6e4272