Reorganized the code so that the various modules are all compiled into a single DSO/DLL called PyJuggler.{dll,so,dylib}.  On Windows, the runtime loader gets confused when it finds two DLLs with the same name, so we cannot have individual DLLs such as gadget.dll and vpr.dll in a PyJuggler directory.  Instead, everything is now in the single PyJuggler module, and each sub-module is (artificially) scoped within that module.  I'm not entirely thrilled with this "solution," but it seems to work and requires no changes to existing Python applications.
As part of this, jccl.cpp has been broken up into per-class .cpp files
in src/jccl.  This was necessary so that the PyJuggler module init function
could call the individual registration functions for the JCCL module.

Also, this change resurrects src/PyJuggler.cpp.  I'm hoping CVS will deal
with this correctly.

Prior to this change, the source tree was tagged with the symbol
PRE_SINGLE_MODULE.  After this commit, it will be tagged again with the
symbol POST_SINGLE_MODULE.

Bumped version to 0.7.1.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/PyJuggler@13071 08b38cba-cd3b-11de-854e-f91c5b6e4272