Exposed more of the JCCL C++ API to Python.  This gives access to most of the JCCL classes that anyone would need for doing basic (and slightly beyond basic) operations with JCCL.
Bumped version to 0.13.0.

NOTE: I have purposely avoided including a lot of the jccl::ConfigManager
interface that relates to manipulation of the pending and active lists.
Unless someone really needs their hands in that stuff, I don't think it is
generally useful.  It is certainly not worth bloating PyJuggler.jccl at
this time.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/PyJuggler@16089 08b38cba-cd3b-11de-854e-f91c5b6e4272