Put the foundation in place to take advantage of the compiler warning levels feature offered by Doozer++.  Right now, additional warnings are disabled ($(WARNS) defaults to 0) because Boost.Python from Boost 1.33 issues far too many warnings.  They would get in the way of trying to find compiler warnings from PyJuggler code.
This change also makes use of @CXXFLAGS@ from the configure script.  I am
not sure if I had a reason for not using this variable before.  I have only
tested this change on Mac OS X 10.4, and using @CXXFLAGS@ there does cause
-fno-common to be added to the compiler line on that platform.  This does
not seem to be introducing any new run-time bugs, so I think that this
change should be safe.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/PyJuggler@17722 08b38cba-cd3b-11de-854e-f91c5b6e4272