Merged extensions from HEAD that take advantage of the compiler warning levels feature offered by Doozer++.  For the pyjutil library, $(WARNS) is set to 2, which is useful, but the extension modules have $(WARNS) set to 0 to disable the expanded warning output.  This is because Boost.Python (at least from Boost 1.33) issues far too many warnings.
This change also makes use of @CXXFLAGS@ from the configure script.  I am
not sure if I had a reason for not using this variable before.  I have only
tested this change on Mac OS X 10.4, and using @CXXFLAGS@ there does cause
-fno-common to be added to the compiler line on that platform.  This does
not seem to be introducing any new run-time bugs, so I think that this
change should be safe.

The specific revisions merged are as follows:

   PyJuggler/pyjutil/Makefile.in: 1.13
   PyJuggler/src/common-targets.mk.in: 1.7
   PyJuggler/src/common.mk.in: 1.5


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/PyJuggler@17727 08b38cba-cd3b-11de-854e-f91c5b6e4272