Reformatted the code to use the Juggler coding standard (with adjustments to acocunt for Boost.Python-isms).  This is largely a whitespace change since Pyste originally generated these files using four spaces of indentation rather than three.  There are no functional changes associated with this revision.
Up until this time, I had been resisting reformatting the code in case I
ever needed to run Pyste again to regenerate the PyJuggler code, but now I
doubt that I will need Pyste again.  It was fine for laying the groundwork,
but the PyJuggler code has been modified and extended in ways that are far
beyond what Pyste can deal with.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/PyJuggler@17730 08b38cba-cd3b-11de-854e-f91c5b6e4272