Make use of "new-style" Boost.Python polymorphism, a feature introduced in Boost 1.32.0.  According to David Abrahams, new-style polymorphism "performs better and has fewer problems in almost every way than the old one."  He also stated that it fixes a memory leak.
This means that Boost 1.32.0 or newer is now required for PyJuggler 1.1.
Furthermore, while this change is completely transparent to Python
programmers using PyJuggler, it represents a major change in the run-time
behavior of PyJuggler with respect to polymorphic types.  Therefore, this
change will not be merged to the 1.0 release branch.

Because of the significance of this change, I tagged the PyJuggler CVS HEAD
branch with the tag PRE_NEW_STYLE_POLYMORPHISM to allow this change to be
backed out easily should that become necessary.  So far, I have only tested
this change on Mac OS X 10.4, but I expect that the extant compiler
requirements for PyJuggler will ensure that the code continues to compile
on Linux, Windows, and FreeBSD.

Finally, it is worth noting that I made these changes over a week ago.
Every change I have made to PyJuggler during the last week has been the
result of having to review a significant portion of the Pyjuggler source
to make the switch to new-style polymorphism.  Since then, I have been
moving the PyJuggler source forward to this point to be able to make this
commit.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/PyJuggler@17749 08b38cba-cd3b-11de-854e-f91c5b6e4272